Finish deserializing index and proof

+++ eth_verifier/lib/VerifierIndex.sol
@@ -16,12 +16,12 @@ struct VerifierIndex {
     uint64 zk_rows;
     URS urs;
     // domain
-    uint256 domain_size;
+    uint64 domain_size;
     Scalar.FE domain_gen;
     /// The mapping between powers of alpha and constraints
     Alphas powers_of_alpha;
     // wire shift coordinates
-    Scalar.FE[7] shift;  // TODO: use Consants.PERMUTS
+    Scalar.FE[7] shift; // TODO: use Consants.PERMUTS
     /// domain offset for zero-knowledge
     Scalar.FE w;
 }
